# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Netwrk API Documentation",
        "description": "API documentation for Netwrk platform - User registration and management",
        "contact": {
            "email": "support@netwrk.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/upload": {
            "post": {
                "tags": [
                    "upload"
                ],
                "summary": "Upload a file to S3",
                "operationId": "9795124f6acf6befebd02b4376d24e97",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "folder": {
                                        "description": "Folder to upload to (optional, defaults to 'uploads')",
                                        "type": "string",
                                        "example": "avatars"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "https://bucket.s3.amazonaws.com/uploads/file.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid file upload",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No file provided"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The file field is required"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/multiple": {
            "post": {
                "tags": [
                    "upload"
                ],
                "summary": "Upload multiple files to S3",
                "operationId": "9256f78f3d1bec5d802ab3e54b260155",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "files[]": {
                                        "description": "Files to upload",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "folder": {
                                        "description": "Folder to upload to (optional, defaults to 'uploads')",
                                        "type": "string",
                                        "example": "documents"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Files uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Files uploaded successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_files": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "successful_uploads": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "failed_uploads": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "uploaded": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "errors": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/delete": {
            "delete": {
                "tags": [
                    "upload"
                ],
                "summary": "Delete a file from S3",
                "operationId": "ff0b2a4b6bbe193a7e7fbf354c34ac3d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "description": "S3 URL of the file to delete",
                                        "type": "string",
                                        "example": "https://bucket.s3.amazonaws.com/uploads/file.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "File deleted successfully"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to delete file",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete file"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/upload/presigned-url": {
            "post": {
                "tags": [
                    "upload"
                ],
                "summary": "Generate a presigned URL for direct upload",
                "operationId": "b6c2cda865969173ed18ff664bf83890",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filename": {
                                        "description": "Original filename",
                                        "type": "string",
                                        "example": "document.pdf"
                                    },
                                    "folder": {
                                        "description": "Folder to upload to (optional, defaults to 'uploads')",
                                        "type": "string",
                                        "example": "documents"
                                    },
                                    "expiration": {
                                        "description": "URL expiration time in seconds (optional, defaults to 3600)",
                                        "type": "integer",
                                        "example": 3600
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Presigned URL generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Presigned URL generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "presigned_url": {
                                                    "type": "string",
                                                    "example": "https://bucket.s3.amazonaws.com/uploads/file.jpg?signature=..."
                                                },
                                                "file_path": {
                                                    "type": "string",
                                                    "example": "uploads/uuid-filename.jpg"
                                                },
                                                "filename": {
                                                    "type": "string",
                                                    "example": "uuid-filename.jpg"
                                                },
                                                "expires_at": {
                                                    "type": "string",
                                                    "example": "2024-01-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}